web1/networking-websockets/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
flask-sock==0.5.0
PyJWT
